requirements.txt
diffusers==0.30.3
accelerate==0.31.0
transformers==4.41.2
einops==0.8.0
decord==0.6.0
ruamel.yaml
hydra-core
fire
pytorch-lightning==2.0.1
kornia==0.6.9
clip@git+https://github.com/openai/CLIP.git
openai-clip
xformers==0.0.29
plyfile
git+https://github.com/ashawkey/kiuikit.git
torchgeometry
diso
scikit-image==0.24.0
onnxruntime
opencv-python
open3d
libigl
ninja
rich
open_clip_torch
rembg
onnxruntime
numpy==1.26.4
imageio==2.19.3
imageio-ffmpeg==0.4.7
av
imageio